322改动
1、登录页面新增重置按钮和输入框相关提示
2、修改密码相关提示信息进行修改和增加电话号码的验证
3、振耀UDP协议功能的修改
4、管理员进行清箱后用户还能进行取件的bug进行修复
5、修改密码的问题

+++ zmailserver/src/main/java/com/zy/zmail/server/north/zyudp/ZyudpReceiver.java
@@ -77,7 +77,7 @@ public class ZyudpReceiver implements Runnable {
     }
 
     public static byte[] extractPacket(byte[] data)  {
-        byte startFlag1 = ZyudpCommand.START_FLAG & 0xFF;
+        byte startFlag1 = (byte)(ZyudpCommand.START_FLAG & 0xFF);
         byte startFlag2 = (byte)(ZyudpCommand.START_REQUEST & 0xFF);
